+++ CHANGES.md
@@ -1,5 +1,14 @@
 # Unreleased (TBD)
 
+* add `reproject` method for `ImageData` objects
+
+    ```python
+    from rio_tiler.models import ImageData
+
+    img = ImageData(numpy.zeros((3, 256, 256), crs=CRS.from_epsg(4326), dtype="uint8"))
+    img.reproject("epsg:3857")
+    ```
+
 * add `indexes` parameter for `XarrayReader` methods. As for Rasterio, the indexes values start at `1`.
 
     ```python